old/requirements.txt
librosa~=0.10.1
numpy~=1.26.3
matplotlib~=3.8.2
scipy~=1.8.0
ffmpeg-python~=0.2.0
PySide6~=6.6.1
pyinstaller~=6.3.0
moviepy~=1.0.3